feat(uploads): Adding Upload Cover Menu

+++ be/src/api/users/employees/admins/handler.js
@@ -1,8 +1,9 @@
 const autoBind = require('auto-bind');
 
 class AdminsHandler {
-  constructor(service, validator) {
+  constructor(service, superAdminsService, validator) {
     this._service = service;
+    this._superAdminsService = superAdminsService;
     this._validator = validator;
 
     autoBind(this);
@@ -10,8 +11,10 @@ class AdminsHandler {
 
   async postAdminsHandler(request, h) {
     this._validator.validatePostAdminPayload(request.payload);
-    const { employeeId } = request.payload;
+    const { id } = request.auth.credentials;
+    await this._superAdminsService.validateSuperAdmin({ id });
 
+    const { employeeId } = request.payload;
     const adminId = await this._service.addAdmins(employeeId);
     const response = h.response({
       status: 'success',
@@ -25,6 +28,9 @@ class AdminsHandler {
   }
 
   async deleteAdminByIdHandler(request, h) {
+    const { id } = request.auth.credentials;
+    await this._superAdminsService.validateSuperAdmin({ id });
+
     const { id: adminId } = request.params;
     await this._service.deleteAdminById(adminId);
     const response = h.response({
