+++ be/src/api/users/employees/chefs/handler.js
@@ -1,8 +1,11 @@
 const autoBind = require('auto-bind');
+const NotFoundError = require('../../../../exceptions/client/NotFoundError');
 
 class ChefsHandler {
-  constructor(service, validator) {
+  constructor(service, adminsService, superAdminsService, validator) {
     this._service = service;
+    this._adminsService = adminsService;
+    this._superAdminsService = superAdminsService;
     this._validator = validator;
 
     autoBind(this);
@@ -10,6 +13,9 @@ class ChefsHandler {
 
   async postChefsHandler(request, h) {
     this._validator.validatePostChefsPayload(request.payload);
+    const { id } = request.auth.credentials;
+
+    await this._superAdminsService.validateSuperAdmin({ id });
     const chefId = await this._service.addChefs(request.payload);
     const response = h.response({
       status: 'success',
@@ -23,6 +29,9 @@ class ChefsHandler {
   }
 
   async getChefsHandler(request, h) {
+    const { id } = request.auth.credentials;
+    await this.verifyAccess({ id });
+
     const chefs = await this._service.getChefs();
     const response = h.response({
       status: 'success',
@@ -34,27 +43,47 @@ class ChefsHandler {
     return response;
   }
 
-  async getChefByIdHandler(request, h) {
+  async getChefByIdHandler(request) {
+    const { id } = request.auth.credentials;
+    await this.verifyAccess({ id });
+
     const { id: chefId } = request.params;
     const chef = await this._service.getChefById(chefId);
-    const response = h.response({
+    return {
       status: 'success',
       message: 'Berhasil mendapatkan chef.',
       data: {
         chef,
       },
-    });
-    return response;
+    };
   }
 
   async deleteChefByIdHandler(request, h) {
+    const { id } = request.auth.credentials;
+    await this._superAdminsService.validateSuperAdmin({ id });
+
     const { id: chefId } = request.params;
     await this._service.deleteChefById(chefId);
-    const response = h.response({
+    return {
       status: 'success',
       message: 'Berhasil menghapus chef.',
-    });
-    return response;
+    };
+  }
+
+  async verifyAccess({ id }) {
+    try {
+      await this._adminsService.verifyAdmin({ id });
+    } catch (error) {
+      if (error instanceof NotFoundError) {
+        try {
+          await this._superAdminsService.validateSuperAdmin({ id });
+        } catch {
+          throw error;
+        }
+      } else {
+        throw error;
+      }
+    }
   }
 }
